Split CapabilityManager into Migrator + RoleRegistrar (#563 Sprint 2)
Break the 1359-line CapabilityManager god-class into three
single-responsibility classes under FreeFormCertificate\UserDashboard\:

- CapabilityMigrator — the one-shot, option-flagged migrations (legacy
  pre-6.2.0 renames, the 6.9.0 taxonomy renames, and the delete/export/
  import/reasons grant back-fills). These run once per install via the
  Loader and don't belong in the live manager.
- RoleRegistrar — the FFC role lifecycle (register/upgrade/remove the
  ffc_user, recruitment-manager, and per-module roles; role relabeling).
- CapabilityManager keeps the runtime surface: capability granting, access
  checks, the *_CAPABILITIES / CONTEXT_* registry, and module_roles_definition()
  (now public so RoleRegistrar can read it). Drops ~1359 -> ~673 lines.

All call sites repointed (loader, activator, user-manager, recruitment-loader,
role-capability-editor) including the wp_roles_init string callback. Tests
repointed to the new classes.

Behavior-preserving: the CapabilityCatalog <-> get_all_capabilities()
invariant holds; full capability/role suites + the complete unit suite green.
PHPStan L8 + WPCS clean. CHANGELOG + CLAUDE.md updated.
