+++ includes/user-dashboard/class-ffc-capability-manager.php
@@ -38,7 +38,7 @@ class CapabilityManager {
 	 * context — candidates rely on the `ffc_user` role's baseline `read` cap
 	 * to access their dashboard "Minhas Convocações" section. The
 	 * `ffc_manage_recruitment` admin cap is registered separately on plugin
-	 * activation (see {@see self::register_recruitment_manager_role()}).
+	 * activation (see {@see RoleRegistrar::register_recruitment_manager_role()}).
 	 *
 	 * @since 6.0.0
 	 */
@@ -58,151 +58,6 @@ class CapabilityManager {
 		'ffc_download_own_certificates',
 		'ffc_view_own_certificate_history',
 	);
-
-	/**
-	 * Map of legacy (pre-6.2.0) cap names → new namespaced names. Consumed
-	 * by the migration helper {@see self::migrate_legacy_certificate_caps()}
-	 * and by `Loader::ensure_legacy_caps_renamed()` on `plugins_loaded`.
-	 *
-	 * @since 6.2.0
-	 * @return array<string, string>
-	 */
-	public static function legacy_cap_renames(): array {
-		return array(
-			'view_own_certificates'     => 'ffc_view_own_certificates',
-			'download_own_certificates' => 'ffc_download_own_certificates',
-			'view_certificate_history'  => 'ffc_view_own_certificate_history',
-		);
-	}
-
-	/**
-	 * Idempotent migration: rewrites every legacy cap grant on every user
-	 * (user-meta `add_cap(true)`) and on the `ffc_user` role to the new
-	 * `ffc_*` namespace.
-	 *
-	 * Strategy: for each `legacy => new` pair,
-	 *   1. iterate every user that has the legacy cap, add the new cap, drop
-	 *      the legacy cap;
-	 *   2. on the `ffc_user` role, if the legacy cap exists, add the new
-	 *      cap with the same boolean value and remove the legacy cap.
-	 *
-	 * Run once per FFC version bump via `Loader::ensure_legacy_caps_renamed()`.
-	 *
-	 * @since 6.2.0
-	 * @return array<string, int> Per-rename count of users migrated.
-	 */
-	public static function migrate_legacy_certificate_caps(): array {
-		$counts = array();
-		foreach ( self::legacy_cap_renames() as $legacy => $renamed ) {
-			$counts[ $legacy ] = 0;
-
-			// 1. User-meta grants. WP_User_Query has no direct `cap` filter,
-			// so we iterate over user IDs and check each — this runs once per
-			// version bump so the cost is acceptable.
-			$users = get_users( array( 'fields' => 'ID' ) );
-			foreach ( $users as $user_id ) {
-				$user = get_userdata( (int) $user_id );
-				if ( ! $user ) {
-					continue;
-				}
-				if ( isset( $user->caps[ $legacy ] ) && true === $user->caps[ $legacy ] ) {
-					$user->add_cap( $renamed, true );
-					$user->remove_cap( $legacy );
-					++$counts[ $legacy ];
-				}
-			}
-
-			// 2. ffc_user role definition.
-			$role = get_role( 'ffc_user' );
-			if ( $role && isset( $role->capabilities[ $legacy ] ) ) {
-				$value = (bool) $role->capabilities[ $legacy ];
-				$role->add_cap( $renamed, $value );
-				$role->remove_cap( $legacy );
-			}
-		}
-		return $counts;
-	}
-
-	/**
-	 * Taxonomy rename map (old => new) for the plugin-wide capability naming
-	 * standard `ffc_<action>_[own_]<domain>[_<qualifier>]`.
-	 *
-	 * ⚠ The `ffc_view_self_scheduling => ffc_view_own_appointments` entry
-	 * **reverses** the historical 4.5.0 migration
-	 * ({@see \FreeFormCertificate\Migrations\MigrationRenameCapabilities},
-	 * which mapped `ffc_view_own_appointments => ffc_view_self_scheduling`).
-	 * It is applied here under a **new** option flag so the two never collide.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, string>
-	 */
-	public static function taxonomy_cap_renames(): array {
-		return array(
-			'ffc_view_certificate_history'    => 'ffc_view_own_certificate_history',
-			'ffc_view_self_scheduling'        => 'ffc_view_own_appointments',
-			'ffc_view_audience_bookings'      => 'ffc_view_own_audience_bookings',
-			'ffc_book_appointments'           => 'ffc_book_own_appointments',
-			'ffc_manage_self_scheduling'      => 'ffc_manage_appointments',
-			'ffc_certificate_update'          => 'ffc_edit_certificates',
-			'ffc_manage_user_custom_fields'   => 'ffc_manage_custom_fields',
-			'ffc_import_recruitment_csv'      => 'ffc_import_recruitment',
-			'ffc_call_recruitment_candidates' => 'ffc_call_recruitment',
-			'ffc_read_forms_api'              => 'ffc_view_forms_api',
-		);
-	}
-
-	/**
-	 * Idempotent migration that rewrites every taxonomy-renamed cap grant on
-	 * (1) every user's user-meta caps and (2) every role definition (including
-	 * `administrator` and the FFC roles), preserving the boolean value.
-	 *
-	 * Runs once per install via {@see \FreeFormCertificate\Loader} on
-	 * `plugins_loaded`, flagged by a dedicated option so it never re-runs and
-	 * never collides with the 4.5.0 rename migration.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, int> Per-rename count of users migrated.
-	 */
-	public static function migrate_taxonomy_renames(): array {
-		$renames = self::taxonomy_cap_renames();
-		$counts  = array();
-
-		// 1. User-meta grants.
-		$users = get_users( array( 'fields' => 'ID' ) );
-		foreach ( $renames as $old => $new ) {
-			$counts[ $old ] = 0;
-			foreach ( $users as $user_id ) {
-				$user = get_userdata( (int) $user_id );
-				if ( ! $user ) {
-					continue;
-				}
-				if ( isset( $user->caps[ $old ] ) ) {
-					$value = (bool) $user->caps[ $old ];
-					$user->add_cap( $new, $value );
-					$user->remove_cap( $old );
-					++$counts[ $old ];
-				}
-			}
-		}
-
-		// 2. Role definitions — administrator + every FFC/custom role.
-		$wp_roles = wp_roles();
-		foreach ( array_keys( $wp_roles->roles ) as $role_slug ) {
-			$role = get_role( $role_slug );
-			if ( ! $role ) {
-				continue;
-			}
-			foreach ( $renames as $old => $new ) {
-				if ( isset( $role->capabilities[ $old ] ) ) {
-					$value = (bool) $role->capabilities[ $old ];
-					$role->add_cap( $new, $value );
-					$role->remove_cap( $old );
-				}
-			}
-		}
-
-		return $counts;
-	}
 
 	/**
 	 * All appointment-related capabilities
@@ -343,304 +198,6 @@ class CapabilityManager {
 		// behavior on upgrade. See `import_cap_grant_map()`.
 		'ffc_import_audiences',
 	);
-
-	/**
-	 * Map of `manage` cap => the `delete` cap it seeds on upgrade (GAP E).
-	 *
-	 * The one-shot migration {@see self::migrate_delete_caps_grant()} grants the
-	 * value cap to every user/role that already holds the key cap, so existing
-	 * managers keep their delete ability after the destructive tier is split out
-	 * of `manage`. To take delete away from a manager, remove the delete cap
-	 * from that role/user after the migration has run.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, string>
-	 */
-	public static function delete_cap_grant_map(): array {
-		return array(
-			'ffc_manage_certificates'   => 'ffc_delete_certificates',
-			'ffc_manage_appointments'   => 'ffc_delete_appointments',
-			'ffc_manage_audiences'      => 'ffc_delete_audiences',
-			'ffc_manage_reregistration' => 'ffc_delete_reregistration',
-			'ffc_manage_custom_fields'  => 'ffc_delete_custom_fields',
-			'ffc_manage_recruitment'    => 'ffc_delete_recruitment',
-			'ffc_manage_url_shortener'  => 'ffc_delete_url_shortener',
-		);
-	}
-
-	/**
-	 * Idempotent migration that seeds each `ffc_delete_<domain>` cap onto every
-	 * user and role that already holds the matching `ffc_manage_<domain>` cap
-	 * (GAP E). Preserves current delete behavior when the destructive tier is
-	 * introduced; never removes a `manage` cap.
-	 *
-	 * Runs once per install via {@see \FreeFormCertificate\Loader} on
-	 * `plugins_loaded`, flagged by the `ffc_delete_caps_granted_v1` option.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, int> Per-delete-cap count of users seeded.
-	 */
-	public static function migrate_delete_caps_grant(): array {
-		$map    = self::delete_cap_grant_map();
-		$counts = array();
-
-		// 1. User-meta grants.
-		$users = get_users( array( 'fields' => 'ID' ) );
-		foreach ( $map as $manage => $delete ) {
-			$counts[ $delete ] = 0;
-			foreach ( $users as $user_id ) {
-				$user = get_userdata( (int) $user_id );
-				if ( ! $user ) {
-					continue;
-				}
-				// Seed only where the manage cap is an explicit grant and the
-				// delete cap isn't already present (idempotent).
-				if ( isset( $user->caps[ $manage ] ) && true === $user->caps[ $manage ]
-					&& ! isset( $user->caps[ $delete ] ) ) {
-					$user->add_cap( $delete, true );
-					++$counts[ $delete ];
-				}
-			}
-		}
-
-		// 2. Role definitions — administrator + every FFC/custom role.
-		$wp_roles = wp_roles();
-		foreach ( array_keys( $wp_roles->roles ) as $role_slug ) {
-			$role = get_role( $role_slug );
-			if ( ! $role ) {
-				continue;
-			}
-			foreach ( $map as $manage => $delete ) {
-				if ( isset( $role->capabilities[ $manage ] ) && true === $role->capabilities[ $manage ]
-					&& ! isset( $role->capabilities[ $delete ] ) ) {
-					$role->add_cap( $delete, true );
-				}
-			}
-		}
-
-		return $counts;
-	}
-
-	/**
-	 * Map of `manage` cap => the `export` cap it seeds on upgrade (GAP G).
-	 *
-	 * The one-shot migration {@see self::migrate_export_caps_grant()} grants the
-	 * value cap to every user/role that already holds the key cap, so existing
-	 * managers keep their bulk-export ability after the export tier is split out
-	 * of `manage`. To take export away from a manager, remove the export cap from
-	 * that role/user after the migration has run.
-	 *
-	 * Certificates is intentionally absent: `ffc_export_certificates` predates
-	 * this split and has always been a standalone cap (never granted by
-	 * `ffc_manage_certificates`), so there is nothing to seed for it.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, string>
-	 */
-	public static function export_cap_grant_map(): array {
-		return array(
-			'ffc_manage_appointments'   => 'ffc_export_appointments',
-			'ffc_manage_reregistration' => 'ffc_export_reregistration',
-			'ffc_manage_audiences'      => 'ffc_export_audiences',
-		);
-	}
-
-	/**
-	 * Idempotent migration that seeds each `ffc_export_<domain>` cap onto every
-	 * user and role that already holds the matching `ffc_manage_<domain>` cap
-	 * (GAP G). Preserves current export behavior when the export tier is split
-	 * out of `manage`; never removes a `manage` cap.
-	 *
-	 * Runs once per install via {@see \FreeFormCertificate\Loader} on
-	 * `plugins_loaded`, flagged by the `ffc_export_caps_granted_v1` option.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, int> Per-export-cap count of users seeded.
-	 */
-	public static function migrate_export_caps_grant(): array {
-		$map    = self::export_cap_grant_map();
-		$counts = array();
-
-		// 1. User-meta grants.
-		$users = get_users( array( 'fields' => 'ID' ) );
-		foreach ( $map as $manage => $export ) {
-			$counts[ $export ] = 0;
-			foreach ( $users as $user_id ) {
-				$user = get_userdata( (int) $user_id );
-				if ( ! $user ) {
-					continue;
-				}
-				// Seed only where the manage cap is an explicit grant and the
-				// export cap isn't already present (idempotent).
-				if ( isset( $user->caps[ $manage ] ) && true === $user->caps[ $manage ]
-					&& ! isset( $user->caps[ $export ] ) ) {
-					$user->add_cap( $export, true );
-					++$counts[ $export ];
-				}
-			}
-		}
-
-		// 2. Role definitions — administrator + every FFC/custom role.
-		$wp_roles = wp_roles();
-		foreach ( array_keys( $wp_roles->roles ) as $role_slug ) {
-			$role = get_role( $role_slug );
-			if ( ! $role ) {
-				continue;
-			}
-			foreach ( $map as $manage => $export ) {
-				if ( isset( $role->capabilities[ $manage ] ) && true === $role->capabilities[ $manage ]
-					&& ! isset( $role->capabilities[ $export ] ) ) {
-					$role->add_cap( $export, true );
-				}
-			}
-		}
-
-		return $counts;
-	}
-
-	/**
-	 * Map of `manage` cap => the `import` cap it seeds on upgrade (GAP H).
-	 *
-	 * The one-shot migration {@see self::migrate_import_caps_grant()} grants the
-	 * value cap to every user/role that already holds the key cap, so existing
-	 * managers keep their bulk-import ability after the import tier is enforced
-	 * strictly. Covers the newly-split `ffc_import_audiences` and the
-	 * `ffc_import_recruitment` cap whose umbrella fallback is removed in 6.9.0
-	 * (custom roles relying on `ffc_manage_recruitment` to import keep working).
-	 * To take import away from a manager, remove the import cap afterward.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, string>
-	 */
-	public static function import_cap_grant_map(): array {
-		return array(
-			'ffc_manage_audiences'   => 'ffc_import_audiences',
-			'ffc_manage_recruitment' => 'ffc_import_recruitment',
-		);
-	}
-
-	/**
-	 * Idempotent migration that seeds each `ffc_import_<domain>` cap onto every
-	 * user and role that already holds the matching `ffc_manage_<domain>` cap
-	 * (GAP H). Preserves current import behavior when the import tier is enforced
-	 * strictly; never removes a `manage` cap.
-	 *
-	 * Runs once per install via {@see \FreeFormCertificate\Loader} on
-	 * `plugins_loaded`, flagged by the `ffc_import_caps_granted_v1` option.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, int> Per-import-cap count of users seeded.
-	 */
-	public static function migrate_import_caps_grant(): array {
-		$map    = self::import_cap_grant_map();
-		$counts = array();
-
-		// 1. User-meta grants.
-		$users = get_users( array( 'fields' => 'ID' ) );
-		foreach ( $map as $manage => $import ) {
-			$counts[ $import ] = 0;
-			foreach ( $users as $user_id ) {
-				$user = get_userdata( (int) $user_id );
-				if ( ! $user ) {
-					continue;
-				}
-				// Seed only where the manage cap is an explicit grant and the
-				// import cap isn't already present (idempotent).
-				if ( isset( $user->caps[ $manage ] ) && true === $user->caps[ $manage ]
-					&& ! isset( $user->caps[ $import ] ) ) {
-					$user->add_cap( $import, true );
-					++$counts[ $import ];
-				}
-			}
-		}
-
-		// 2. Role definitions — administrator + every FFC/custom role.
-		$wp_roles = wp_roles();
-		foreach ( array_keys( $wp_roles->roles ) as $role_slug ) {
-			$role = get_role( $role_slug );
-			if ( ! $role ) {
-				continue;
-			}
-			foreach ( $map as $manage => $import ) {
-				if ( isset( $role->capabilities[ $manage ] ) && true === $role->capabilities[ $manage ]
-					&& ! isset( $role->capabilities[ $import ] ) ) {
-					$role->add_cap( $import, true );
-				}
-			}
-		}
-
-		return $counts;
-	}
-
-	/**
-	 * Map of source cap => the recruitment-reasons cap it seeds on upgrade
-	 * (GAP I). Unlike the delete/export/import maps (all keyed on a `manage`
-	 * cap), reasons split into a *pair*: the read tier is seeded from whoever
-	 * could already see the Reasons tab (`ffc_view_recruitment`), and the edit
-	 * tier from whoever could already edit reasons via the umbrella
-	 * (`ffc_manage_recruitment`). This preserves both read and edit access when
-	 * the reasons sub-domain is carved out of the page/umbrella caps.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, string>
-	 */
-	public static function reasons_cap_grant_map(): array {
-		return array(
-			'ffc_view_recruitment'   => 'ffc_view_recruitment_reasons',
-			'ffc_manage_recruitment' => 'ffc_manage_recruitment_reasons',
-		);
-	}
-
-	/**
-	 * Idempotent migration that seeds the recruitment-reasons caps onto every
-	 * user/role that already holds the matching source cap (GAP I). Preserves
-	 * current read/edit access when the Reasons tab is moved onto its own strict
-	 * 3-state tier; never removes a source cap.
-	 *
-	 * Runs once per install via {@see \FreeFormCertificate\Loader} on
-	 * `plugins_loaded`, flagged by the `ffc_reasons_caps_wired_v1` option.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, int> Per-reasons-cap count of users seeded.
-	 */
-	public static function migrate_reasons_caps_grant(): array {
-		$map    = self::reasons_cap_grant_map();
-		$counts = array();
-
-		// 1. User-meta grants.
-		$users = get_users( array( 'fields' => 'ID' ) );
-		foreach ( $map as $source => $reasons_cap ) {
-			$counts[ $reasons_cap ] = 0;
-			foreach ( $users as $user_id ) {
-				$user = get_userdata( (int) $user_id );
-				if ( ! $user ) {
-					continue;
-				}
-				if ( isset( $user->caps[ $source ] ) && true === $user->caps[ $source ]
-					&& ! isset( $user->caps[ $reasons_cap ] ) ) {
-					$user->add_cap( $reasons_cap, true );
-					++$counts[ $reasons_cap ];
-				}
-			}
-		}
-
-		// 2. Role definitions — administrator + every FFC/custom role.
-		$wp_roles = wp_roles();
-		foreach ( array_keys( $wp_roles->roles ) as $role_slug ) {
-			$role = get_role( $role_slug );
-			if ( ! $role ) {
-				continue;
-			}
-			foreach ( $map as $source => $reasons_cap ) {
-				if ( isset( $role->capabilities[ $source ] ) && true === $role->capabilities[ $source ]
-					&& ! isset( $role->capabilities[ $reasons_cap ] ) ) {
-					$role->add_cap( $reasons_cap, true );
-				}
-			}
-		}
-
-		return $counts;
-	}
 
 	/**
 	 * Slug for the dedicated recruitment-manager role.
@@ -1016,117 +573,6 @@ class CapabilityManager {
 	}
 
 	/**
-	 * Register ffc_user role on plugin activation
-	 *
-	 * @return void
-	 */
-	public static function register_role(): void {
-		$existing_role = get_role( 'ffc_user' );
-
-		if ( $existing_role ) {
-			self::upgrade_role( $existing_role );
-			return;
-		}
-
-		// `ffc_user` only owns the WordPress baseline `read` cap. FFC caps are
-		// granted per-user via `grant_*_capabilities()` (user-meta `add_cap(true)`).
-		//
-		// Historically every FFC cap was added here as `=> false` to make the
-		// role's surface explicit, but that breaks multi-role users (e.g. an
-		// administrator who is also an `ffc_user` for their own certificates):
-		// WP's `WP_User::get_role_caps()` merges role capability maps via
-		// `array_merge()`, and an explicit `false` in one role overwrites a
-		// `true` from another. An absent key, by contrast, lets the other
-		// role's `true` survive. See issue #86.
-		add_role(
-			'ffc_user',
-			__( 'FFC User', 'ffcertificate' ),
-			array( 'read' => true )
-		);
-	}
-
-	/**
-	 * Upgrade existing ffc_user role with new capabilities
-	 *
-	 * @since 4.4.0
-	 * @param \WP_Role $role Existing role object.
-	 * @return void
-	 */
-	private static function upgrade_role( \WP_Role $role ): void {
-		// Strip every legacy `=> false` entry for FFC caps so multi-role users
-		// (e.g. administrator + ffc_user) don't have admin-granted caps masked
-		// by ffc_user's explicit denial. New caps are NOT added — absent ≡
-		// false for `current_user_can()` on single-role users, while letting
-		// `array_merge()` preserve `true` from a peer role for multi-role
-		// users. See issue #86 / register_role() for the full rationale.
-		foreach ( self::get_all_capabilities() as $cap ) {
-			if ( isset( $role->capabilities[ $cap ] ) && false === $role->capabilities[ $cap ] ) {
-				$role->remove_cap( $cap );
-			}
-		}
-	}
-
-	/**
-	 * Remove ffc_user role on plugin deactivation
-	 *
-	 * @return void
-	 */
-	public static function remove_role(): void {
-		remove_role( 'ffc_user' );
-	}
-
-	/**
-	 * Register the dedicated recruitment-manager role on plugin activation.
-	 *
-	 * Idempotent: if the role already exists, it is upgraded so any newly
-	 * introduced caps are added without disturbing other caps the admin may
-	 * have manually granted. Granted caps: `read` + `ffc_manage_recruitment`.
-	 *
-	 * Called from {@see \FreeFormCertificate\Activator::activate()}.
-	 *
-	 * @since 6.0.0
-	 * @return void
-	 */
-	public static function register_recruitment_manager_role(): void {
-		$existing_role = get_role( self::RECRUITMENT_MANAGER_ROLE );
-
-		if ( $existing_role ) {
-			// Upgrade path: ensure both caps are present without overwriting
-			// any admin-customized capability map.
-			if ( ! isset( $existing_role->capabilities['read'] ) ) {
-				$existing_role->add_cap( 'read', true );
-			}
-			if ( ! isset( $existing_role->capabilities['ffc_manage_recruitment'] ) ) {
-				$existing_role->add_cap( 'ffc_manage_recruitment', true );
-			}
-			return;
-		}
-
-		add_role(
-			self::RECRUITMENT_MANAGER_ROLE,
-			__( 'Recruitment Manager', 'ffcertificate' ),
-			array(
-				'read'                   => true,
-				'ffc_manage_recruitment' => true,
-			)
-		);
-	}
-
-	/**
-	 * Remove the recruitment-manager role on plugin uninstall.
-	 *
-	 * Called from `uninstall.php` after the cap-stripping loop has cleared
-	 * `ffc_manage_recruitment` from any user that had been granted it
-	 * directly.
-	 *
-	 * @since 6.0.0
-	 * @return void
-	 */
-	public static function remove_recruitment_manager_role(): void {
-		remove_role( self::RECRUITMENT_MANAGER_ROLE );
-	}
-
-	/**
 	 * Definition map for the 6.2.0 module-manager + recruitment-tier roles.
 	 *
 	 * Each entry is a `slug => array{label: string, caps: list<string>}` so
@@ -1145,7 +591,7 @@ class CapabilityManager {
 	 * @since 6.2.0
 	 * @return array<string, array{label: string, caps: list<string>}>
 	 */
-	private static function module_roles_definition(): array {
+	public static function module_roles_definition(): array {
 		return array(
 			// ── Cross-module roles ───────────────────────────────────────
 			// `ffc_administrator` is the aggregator (GAP F): a full FFC admin
@@ -1222,138 +668,5 @@ class CapabilityManager {
 				),
 			),
 		);
-	}
-
-	/**
-	 * Register every 6.2.0 module-manager + recruitment-tier role.
-	 *
-	 * Idempotent. Called on activation AND from `RecruitmentLoader` on
-	 * `plugins_loaded` so in-place plugin updates self-heal — no need
-	 * for a deactivate/reactivate cycle to surface new roles.
-	 *
-	 * Existing roles are upgraded: missing caps are added; extra caps
-	 * an operator manually granted are NOT removed.
-	 *
-	 * @since 6.2.0
-	 * @return void
-	 */
-	public static function register_module_roles(): void {
-		foreach ( self::module_roles_definition() as $slug => $def ) {
-			$existing = get_role( $slug );
-			if ( $existing ) {
-				if ( ! isset( $existing->capabilities['read'] ) ) {
-					$existing->add_cap( 'read', true );
-				}
-				foreach ( $def['caps'] as $cap ) {
-					if ( ! isset( $existing->capabilities[ $cap ] ) ) {
-						$existing->add_cap( $cap, true );
-					}
-				}
-				continue;
-			}
-
-			$caps_map = array( 'read' => true );
-			foreach ( $def['caps'] as $cap ) {
-				$caps_map[ $cap ] = true;
-			}
-			add_role( $slug, $def['label'], $caps_map );
-		}
-
-		// Upgrade the existing `ffc_recruitment_manager` (6.0.0) to the new
-		// tier-2 cap set. Adds the 6.2.0 granular caps without removing the
-		// umbrella `ffc_manage_recruitment` cap that was there from launch.
-		$existing_manager = get_role( self::RECRUITMENT_MANAGER_ROLE );
-		if ( $existing_manager ) {
-			$tier_2_caps = array(
-				'ffc_view_recruitment',
-				'ffc_call_recruitment',
-				'ffc_import_recruitment',
-				'ffc_view_recruitment_pii',
-				'ffc_delete_recruitment',
-				// GAP I: reasons split into their own strict tier. A plain
-				// Recruitment Manager previously edited reasons via the umbrella;
-				// carry both reasons caps explicitly so that keeps working.
-				'ffc_view_recruitment_reasons',
-				'ffc_manage_recruitment_reasons',
-			);
-			foreach ( $tier_2_caps as $cap ) {
-				if ( ! isset( $existing_manager->capabilities[ $cap ] ) ) {
-					$existing_manager->add_cap( $cap, true );
-				}
-			}
-		}
-	}
-
-	/**
-	 * Remove every 6.2.0 module-manager + recruitment-tier role on
-	 * plugin uninstall.
-	 *
-	 * @since 6.2.0
-	 * @return void
-	 */
-	public static function remove_module_roles(): void {
-		foreach ( array_keys( self::module_roles_definition() ) as $slug ) {
-			remove_role( $slug );
-		}
-	}
-
-	/**
-	 * Re-translate FFC role labels at every page load.
-	 *
-	 * WordPress stores role labels verbatim in the `wp_user_roles` option
-	 * at the moment of `add_role()`. If translations weren't loaded yet,
-	 * the English string gets frozen in the database. On subsequent loads,
-	 * `users.php` displays role names via WP's `translate_user_role()`
-	 * helper — but that helper resolves the label against the **default**
-	 * WP textdomain, not the plugin's textdomain. So plugin-provided role
-	 * labels never translate, even after the .po file is loaded.
-	 *
-	 * Hooking `wp_roles_init` (since WP 4.7) lets us mutate the in-memory
-	 * `WP_Roles::$roles` + `WP_Roles::$role_names` arrays after they're
-	 * loaded. We re-resolve every FFC role's label through `__()` against
-	 * the plugin's textdomain, so users.php always shows the user's locale.
-	 *
-	 * Hooked from `Loader::register_ffc_roles_safe()` (init:1) AFTER the
-	 * role registrations themselves, so the freshly-registered labels also
-	 * get re-translated on the very first page load post-upgrade.
-	 *
-	 * @since 6.2.0
-	 * @param \WP_Roles $wp_roles Roles instance.
-	 * @return void
-	 */
-	public static function relabel_ffc_roles( \WP_Roles $wp_roles ): void {
-		$labels = self::ffc_managed_role_labels();
-
-		foreach ( $labels as $slug => $label ) {
-			if ( isset( $wp_roles->roles[ $slug ] ) ) {
-				$wp_roles->roles[ $slug ]['name'] = $label;
-			}
-			if ( isset( $wp_roles->role_names[ $slug ] ) ) {
-				$wp_roles->role_names[ $slug ] = $label;
-			}
-		}
-	}
-
-	/**
-	 * Canonical map of every FFC-managed role slug → display label.
-	 *
-	 * The authoritative set of roles the plugin owns: `ffc_user`, the
-	 * recruitment manager, and the 6.2.0 module/recruitment-tier roles. This
-	 * is the list the role-capability editor (Settings → User Access) is
-	 * allowed to touch — independent of whichever caps a role currently
-	 * carries, so a role whose FFC caps were all unchecked still appears.
-	 *
-	 * @since 6.9.0
-	 * @return array<string, string>
-	 */
-	public static function ffc_managed_role_labels(): array {
-		$labels = array(
-			'ffc_user'                     => __( 'FFC User', 'ffcertificate' ),
-			self::RECRUITMENT_MANAGER_ROLE => __( 'Recruitment Manager', 'ffcertificate' ),
-		);
-		foreach ( self::module_roles_definition() as $slug => $def ) {
-			$labels[ $slug ] = $def['label'];
-		}
-		return $labels;
 	}
 }
